+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- `fullscreen` rule to hide notifications when a fullscreen window is active
+
 ## 1.3.0 - 2018-01-05
 
 ### Added